Fix webGUI port
The Port # [IP]:[PORT] always refers to the container port.  This allows the docker system to see what port (80) is mapped to (8384) and have the URL automatically updated if the user changes the host port.  While yours does work, if the user changes the host port to something else more suitable to them then they also have to switch to advanced view and change the port # in the webGUI